### 2024-09-23 6eb2707
update paper.bib doi
