+++ paper/paper.bib
@@ -4,6 +4,7 @@
   journal      = {Physical Review E},
   volume       = {109},
   pages        = {054306},
+  doi          = {10.1103/physreve.109.054306},
   year         = {2024}
 }
 
@@ -13,6 +14,7 @@
   journal      = {Communications Physics},
   volume       = {6},
   pages        = {148},
+  doi          = {10.1038/s42005-023-01270-5},
   year         = {2023}
 }
 
@@ -22,6 +24,7 @@
   journal      = {Communications Physics},
   volume       = {5},
   pages        = {249},
+  doi          = {10.1038/s42005-022-01029-4},
   year         = {2022}
 }
 
@@ -31,6 +34,7 @@
   journal      = {Physical Review E},
   volume       = {109},
   pages        = {034310},
+  doi          = {10.1103/physreve.109.034310},
   year         = {2024}
 }
 
@@ -40,17 +44,20 @@
   journal      = {Physical Review Research},
   volume       = {6},
   pages        = {033307},
+  doi          = {10.1103/physrevresearch.6.033307},
   year         = {2024}
 }
 
 @misc{kirkley2024fastnonparametricinferencenetwork,
-      title={Fast nonparametric inference of network backbones for graph sparsification}, 
-      author={Alec Kirkley},
-      year={2024},
-      eprint={2409.06417},
-      archivePrefix={arXiv},
-      primaryClass={cs.SI},
-      url={https://arxiv.org/abs/2409.06417}, 
+  title={Fast nonparametric inference of network backbones for graph sparsification}, 
+  author={Alec Kirkley},
+  year={2024},
+  eprint={2409.06417},
+  archivePrefix={arXiv},
+  primaryClass={cs.SI},
+  url={https://arxiv.org/abs/2409.06417}, 
+  doi={10.48550/arXiv.2409.06417},
+
 }
 
 @article{peixoto_graph-tool_2014,
@@ -82,7 +89,8 @@
   title={Networks},
   author={Newman, Mark},
   year={2018},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/oso/9780198805090.001.0001}
 }
 @article{fortunato2010community,
   title={Community detection in graphs},
@@ -91,6 +99,7 @@
   volume={486},
   number={3-5},
   pages={75--174},
+  doi={10.1016/j.physrep.2009.11.002},
   year={2010},
   publisher={Elsevier}
 }
@@ -102,6 +111,7 @@
   number={1},
   pages={6794},
   year={2022},
+  doi={10.1038/s41467-022-34267-9},
   publisher={Nature Publishing Group UK London}
 }
 @article{young2022clustering,
@@ -112,13 +122,15 @@
   number={1},
   pages={014312},
   year={2022},
+  doi={10.1103/physreve.105.014312},
   publisher={APS}
 }
 
 @techreport{hagberg2008exploring,
-  title={Exploring network structure, dynamics, and function using {N}etwork{X}},
+  title={Exploring network structure, dynamics, and function using NetworkX},
   author={Hagberg, Aric and Swart, Pieter J and Schult, Daniel A},
   year={2008},
+  doi={10.25080/tcwv9851},
   institution={Los Alamos National Laboratory (LANL), Los Alamos, NM (United States)}
 }
 
@@ -130,6 +142,7 @@
   number={10},
   pages={1093--1098},
   year={2021},
+  doi={10.1038/s41567-021-01371-4},
   publisher={Nature Publishing Group UK London}
 }
 
@@ -141,6 +154,7 @@
   number={1},
   pages={011010},
   year={2022},
+  doi={10.1103/physrevx.12.011010},
   publisher={APS}
 }
 
[…]
