### 2024-10-02 #7263 @vissarion
Hi @aneeshnaik, @ankurankan, @gchure, @matt-graham, @rafaelorozco would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
