### 2024-10-03 #7312 @editorialbot
[REVIEW]: PANINIpy: Package of Algorithms for Nonparametric Inference with Networks in Python
**Submitting author:** @baiyueh (<a href="http://orcid.org/0009-0007-9787-9726">Baiyue He</a>)
**Repository:** https://github.com/baiyueh/PANINIpy
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.1
