### 2024-10-14 baiyueh/PANINIpy#1 @gchure
JOSS Review: Documentation Requirements
Hi! I'm reviewing [your paper for JOSS](https://github.com/openjournals/joss-reviews/issues/7312) and am opening issues as I make my way through the checklist. This issue pertains to adding materials to the documentation. I'll make another issue for specific comments on extant documentation.

The tutorial notebooks look rather good, but there are some additions to the documentation that are necessary. Could you please add the following as either sections on the landing page or individual pages themselves?

1. Please add a brief summary of the software, and what problem it is designed to solve.
2. Please add a statement of need (as required by JOSS), ideally linking to other packages which can be used to perform network inference.
3. Please add guidelines for contributing to the development of the project. This includes preferred (if any) issue templates and instructions for branch management.
4. Please add installation instructions and a minimal working example for the user to ensure they have a working environment.
