### 2024-10-16 0d4dd8f
adjust landing page, add instllation, contributing pages
