### 2024-10-17 baiyueh/PANINIpy#1 @gchure
Hi @baiyueh, looks great. These changes address all of my comments in this issue. I'll continue to the rest of my review and will open up some more issues.

Thanks for the fast response, and congrats on a sick logo. It looks like it may have been made using generative AI (maybe I'm wrong!). If so, please add a small credit for which model was used to generate it at the bottom of your readme and DOC landing page.
