### 2024-10-20 f2e2f27
add Attribution section
