### 2024-11-02 #7312 @ankurankan
The package appears well-developed and includes a range of examples, making it accessible. My primary feedback is focused on the paper, which in my opinion can benefit from some refinement to ensure that key points come across more clearly.

Specific Feedback:

Installation: Currently, the package does not automatically install its dependencies, nor could I locate a requirements file. Running the example code in the documentation led to errors and based on that I had to manually install numpy, pandas, and scipy. I think these dependencies should get installed automatically.

Automated Tests: I couldn't find any automated tests for the package’s functionality. If I missed these, could you please point them out?

Statement of Need: I think this section can be refined to improve readability and clarity. For example, the last sentences in the first paragraph appear to hint towards the need for non-parametric methods, but this point is not clearly articulated since "non-parametric" isn't mentioned explicitly. Additionally, the next paragraph introduces non-parametric methods, but because the motivation isn't clear in the last paragraph it isn't clear why is it needed. I also think, in the current text, there isn't a clear distinction between the need for non-parametric methods and the need for the software package.

Related Software Packages: The arguments here could be clarified to more effectively illustrate the specific advantages of PANINIpy. For instance, in the comparison with Graph-Tool, it’s not fully clear in which cases PANINIpy should be preferred over Graph-Tool. Based on the text, it seems that there is some overlap, but is it just about the ease of use or there are distinct methods. Clarifying these distinctions would help reinforce the contributions of PANINIpy.
