### 2024-11-04 baiyueh/PANINIpy#3 @gchure
JOSS Review: License Conflict + Requirements File
Hi! I'm reviewing your project for [JOSS](https://github.com/openjournals/joss-reviews/issues/7312) and am opening issues as I progress through the checklist.

First, there is a conflict in licensing for PANINIpy. Your repository [claims](https://github.com/baiyueh/PANINIpy/blob/main/LICENSE.txt) it is licensed under an MIT standard license, however your [`setup.py` file](https://github.com/baiyueh/PANINIpy/blob/f2e2f27d94535aa6382f67017833c91e483bbbb9/setup.py#L12C1-L12C29) assigns an ["Unlicense"](https://choosealicense.com/licenses/unlicense/) which is a legally ambiguous license (and mybe not valid in the yes of JOSS). Please converge on a license and have that unified across the project.

Second, there is no apparent requirements file for the project, leading to installation failures into a sterile venv. Please add requirements either as an external file or within the `setup.py`.
