### 2024-11-04 baiyueh/PANINIpy#4 @gchure
JOSS Review: Automated Testing
Hi! I'm reviewing your paper for [JOSS](https://github.com/openjournals/joss-reviews/issues/7312) and am opening issues as I go through my checklist.

This package does not appear to have any testing, which is a pretty big blocking omission. If you do have tests in your local development, please add and commit them into the repository and set them up for continuous integration.

Happy to discuss this more, but this is a critical issue to address to continue with the review.
