### 2024-11-04 #7312 @gchure
Hi @vissarion, I've made it as far as I can right now in my review. I've opened up issues regarding licensing, documentation, and testing and view those as blocking. Like @ankurankan, I had trouble installing this project due to the lack of clear requirements. Additionally, the package appears to be missing testing writ large, which should be addressed before I continue assessing the functionality claimed in the project.
