### 2024-11-05 ad9c706
add auto-test workflow and add auto-install dependencies
