+++ requirements.txt
@@ -1,0 +1,4 @@
+numpy>=1.24
+pandas>=2.2
+scipy>=1.10
+pytest>=7.3
