+++ PANINIpy/MANIFEST.in
@@ -1,3 +1,4 @@
 graft paper
 include setup.py
 global.exclude *.py[cod]
+exclude Tests/*

[…]
