### 2024-11-05 b595771
add auto-test workflow and add auto-install dependencies
