### 2024-11-05 ac10194
add auto-test workflow and add auto-install dependencies
