+++ .github/workflows/auto-test.yml
@@ -17,5 +17,6 @@ jobs:
           python -m pip install --upgrade pip
           pip install -r requirements.txt
           pip install pytest
+          pip install paninipy
       - name: Run Auto-Tests
         run: pytest Tests/
