### 2024-11-05 2510d5c
add auto-test workflow and add auto-install dependencies
