+++ requirements.txt
@@ -1,4 +1,3 @@
 numpy>=1.24
 pandas>=2.2
 scipy>=1.10
-pytest>=7.3
