### 2024-11-05 ff7bca8
add test instruction on installation page
