### 2024-11-05 a41f15d
add test instruction on installation page
