### 2024-11-05 61438d0
add test instruction on installation page
