### 2024-11-05 #7312 @baiyueh
Hi @ankurankan and @gchure, thank you for your helpful suggestions about the Installation, Automated Tests, Statement of Need and Related Software Packages sections, as well as your time to review the paper. We've addressed the paper-related comments as follows:

(1) We've added the CI for auto-testing, which has been discussed here under the section [Testing and Continuous Integration](https://paninipy.readthedocs.io/en/latest/Papers/installation.html). Besides, all dependencies have been tested under the clean test environment and through series of [workflow](https://github.com/baiyueh/PANINIpy/actions).

(2) We've clarified the motivating first paragraph in question by explicitly mentioning the nonparametric requirement that is central to the package, which allows for the avoidance of ad hoc parameter choices and provides robustness to noise that we mention is important.

(3) We've edited our discussion of existing methods to focus on the novelty and importance of having a separate PANINIpy package. In particular, we've emphasized that the existing packages only have methods for community detection and/or network reconstruction (two network inference tasks which we do not address with PANINIpy), and that PANINIpy has a wide breadth of different unsupervised inference methods (e.g. hub identification, network population clustering, etc) whose results can be compared on a universal scale (data compression in bits) to find parsimonious summaries of networks from multiple perspectives. We have tried to emphasize in the revision that there is actually little overlap with these other packages in terms of the inference tasks being addressed, which can enhance the justification for the need for the package.

(4) We have combined the "Statement of Need" and "Related Software Packages" sections as we feel the argument for the package's novelty and importance can be made more concise this way. But we are happy to separate them if this is desired.

Let us know if you have further concerns regarding the software, paper or documentation. Happy to help.
