### 2024-11-05 baiyueh/PANINIpy#3 @baiyueh
Hi @gchure, thanks for bringing that up, have fixed to MIT license same as the github repo. Have added both requirments.txt and include the dependencies automatically while pip install.
