### 2024-11-05 baiyueh/PANINIpy#4 @baiyueh
Hi @gchure, the test section with pytest has been added for both insturction for local [Testing and Continuous Integration](https://paninipy.readthedocs.io/en/latest/Papers/installation.html), and CI flow on GitHub [workflow](https://github.com/baiyueh/PANINIpy/actions). Please let us know if this could resolve your issue, thanks!
