### 2024-11-05 baiyueh/PANINIpy#4 @gchure
Great, thank you! I cloned the repository to run the tests and encountered some issues with installation. I was able to resolve them by fixing some issues in your `setup.py`.

To run the tests, I created a virtual environment on my local machine running python 3.11 and installed paninipy via `pip install -e .`. Doing so installed the package, but it could not be imported as `import paninipy`. This is because your `setup.py` file says the package name is `paninipy`, but your module directlry is `PANINIpy`. Running `import PANINIpy` ran without issue.

I would suggest changing the root folder to be `paninipy` (all lower caps) to be consistent with the documentation and tests.

Also, there's a small typo in your `setup.py`. The field `project_urls` is mistyped as `projects_urls`.
