### 2024-11-05 baiyueh/PANINIpy#4 @gchure
Otherwise, the tests all seem to run, so once the naming convention is addressed you can close this issue.
