### 2024-11-06 38d5734
change the test node python version to 3.11
