### 2024-11-06 baiyueh/PANINIpy#4 @baiyueh
Hi @gchure, I have run the test node with Python 3.11 without any issues, and I believe the package naming issue was resolved previously. Could you please check the version of paninipy by running `pip show paninipy`? If it is below 1.0, you will need to either reinstall it using `pip uninstall paninipy --yes` followed by `pip install paninipy`, or simply update the current version with `pip install --upgrade paninipy` to resolve the issue.

Thanks, just fixed the typo thing in `setup.py`.
