### 2024-11-08 baiyueh/PANINIpy#4 @baiyueh
> Okay, interesting. I'm still getting it to fail locally when I install paninipy from source (cloning the repo and then running `pip install -e)`. I get the same result if I do `pip install paninipy`
>
> Here's the output of `pip show paninipy`:
>
> ```
> Name: paninipy
> Version: 1.1
> Summary: Package of Algorithms for Nonparametric Inference with Networks in Python
> Home-page: https://paninipy.readthedocs.io/en/latest/index.html
> Author: Baiyue He
> Author-email: baiyue.he@connect.hku.hk
> License: The MIT License
> Location: /Users/gchure/Desktop/paninipy
> Requires: numpy, pandas, scipy
> Required-by:
> ```
>
> The result of `pip freeze` in my venv yields the following:
>
> ```
> numpy==2.1.3
> pandas==2.2.3
> paninipy==1.1
> python-dateutil==2.9.0.post0
> pytz==2024.2
> scipy==1.14.1
> six==1.16.0
> tzdata==2024.2
> ```
>
> With this (and having pulled the most recent commit of the repo), running `pytest` from the root directory fails with the following: [paninipy_failure.txt](https://github.com/user-attachments/files/17668379/paninipy_failure.txt)
>
> This is fixed by simply running `mv PANINIpy paninipy && pytest` with the following output: [paninipy_success.txt](https://github.com/user-attachments/files/17668392/paninipy_success.txt)
>
> My local environment is MacOS with OS Sonoma 14.6.1.
