### 2024-11-08 baiyueh/PANINIpy#4 @baiyueh
Hi @gchure, thanks for providing a detailed description of the error occurrence, which helped me locate the root cause.

It turns out that my local environment is on Windows, where the file system is not case-sensitive, causing git to ignore renames that only change casing. The same applied to my local test runs. The issue it that the `pytest` imports specifically from the `paninipy` folder, which caused a mismatch due to case sensitivity. This is unrelated to the installation method (`pip install paninipy` or `pip install -e .`).

I have made the necessary adjustments to enforce lowercase usage. Here is the updated flow I followed for your reference:
[test_env_flow.txt](https://github.com/user-attachments/files/17673031/test_env_flow.txt)

Please check it out and let me know if the issue is resolved. Thank you!
