### 2024-11-08 #7312 @gchure
Hi @vissarion, I closed my last remaining issue with PANINIpy and have checked the last box on my review checklist. I think it's ready to go! Thanks @baiyueh for the rapid responses on my issues, and sorry this review took longer than I anticipated.
