### 2024-11-11 #7312 @ankurankan
@baiyueh Thanks for the changes, I think the paper reads much better now. A very minor suggestion about the current modules section: The citations are currently within brackets which look a bit weird in sentences. I don't know if there is a way to add citations without brackets? Else, you could potentially rewrite these sentences in a bit more compact form like: "Methods (A. Kirkley, 2024b) for identifying MDL-optimal temporally contiguous partitions ...". This should also make the paper fit in 2 pages.

@vissarion I have updated my checklist and I also think this is ready to go.
