### 2024-11-12 #7312 @aleckirkley
@vissarion @ankurankan @gchure Thank you all for engaging us in the JOSS review process and helping us to improve the package! We hope it can serve as a useful tool for people across disciplines using networks in their research.
