### 2024-11-12 baiyueh/PANINIpy#5 @vissarion
JOSS review
Α minor comment on the draft:

- line 9, please fix reference  [Barabási 2016]
