### 2024-11-12 #7312 @vissarion
Thanks, @baiyueh I have a minor comment, see https://github.com/baiyueh/PANINIpy/issues/5
