### 2024-11-12 4242536
paper citation adjust
