### 2024-11-12 ee44ebb
add doi and adjust setup
