+++ setup.py
@@ -2,13 +2,13 @@ from setuptools import setup, find_packages, Extension
 
 setup(
     name='paninipy',
-    version='1.2',
+    version='1.3',
     description='Package of Algorithms for Nonparametric Inference with Networks in Python',
     long_description=open('README.md', encoding='utf-8').read(),
     long_description_content_type='text/markdown',
     url='https://paninipy.readthedocs.io/en/latest/index.html',
-    author='Baiyue He',
-    author_email='baiyue.he@connect.hku.hk',
+    author='Baiyue He, Alec Kirkley',
+    author_email='baiyue.he@connect.hku.hk, akirkley@hku.hk',
     license='The MIT License',
     project_urls={
         "Documentation": "https://paninipy.readthedocs.io/en/latest/index.html",
