### 2024-11-12 2627617
remove the draft pdf workflow
