### 2024-11-12 #7312 @baiyueh
Hi @vissarion, have addressed the minor issue with the citation.

The published the release of PANINIpy-v1.0.1 with DOI: 10.5281/zenodo.14100356. Do let us know if there is anything missing out there.
