### 2024-11-14 #7312 @crvernon
:wave: @baiyueh - I just need you to address the following before I accept this one for publication:

In the archive:

- [ ] The license specified in your Zenodo archive should be the same as the one in your code repository.  Please edit the metadata in your Zenodo archive to fix this.

In the paper:

- [ ] Please remove the formatting (the bold designations) in the title of your paper.  Otherwise it will not show up correctly in JOSS.

That's all.  This paper was very clean.  Thank you for that!  Let me know when you have made these changes.
